Move noise learner decoder
